# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stellar Agent Trust Fabric API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: Browse service registry sorted by reputation + price
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Service list
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
                  total:
                    type: integer
                required: [services, total]

    post:
      operationId: registerService
      tags: [services]
      summary: Register a new x402-protected service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterServiceBody"
      responses:
        "201":
          description: Service registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{serviceId}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service details
      parameters:
        - in: path
          name: serviceId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /services/paid/summarize:
    post:
      operationId: paidSummarize
      tags: [services]
      summary: x402-protected AI summarizer endpoint (requires payment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                agentId:
                  type: string
                sessionToken:
                  type: string
              required: [text, agentId]
      responses:
        "200":
          description: Summary result
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                  wordCount:
                    type: integer
                  paymentRecorded:
                    type: boolean
                  paymentId:
                    type: string
                required: [summary, wordCount, paymentRecorded]
        "402":
          description: Payment Required — x402 flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequiredResponse"

  # ── Demo agent ───────────────────────────────────────────────────────────────
...
  /stellar/account/{address}/balance:
    get:
      operationId: getStellarBalance
      tags: [stellar]
      summary: Get XLM and USDC balance for a Stellar address
      parameters:
        - in: path
          name: address
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Account balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StellarBalanceResponse"
        "400":
          description: Invalid address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stellar/payment/build:
    post:
      operationId: buildStellarPayment
      tags: [stellar]
      summary: Build an MPP-style split payment XDR (90% service, 10% protocol fee)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildPaymentBody"
      responses:
        "200":
          description: Unsigned XDR transaction + MPP split breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildPaymentResponse"
  /stellar/payment/submit:
    post:
      operationId: submitStellarPayment
      tags: [stellar]
      summary: Submit a signed XDR transaction to Stellar Testnet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                xdr:
                  type: string
              required: [xdr]
      responses:
        "200":
          description: Transaction submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitPaymentResponse"
  /stellar/payment/verify/{txHash}:
    get:
      operationId: verifyStellarPayment
      tags: [stellar]
      summary: Verify a Stellar USDC payment transaction on Horizon
      parameters:
        - in: path
          name: txHash
          required: true
          schema: { type: string }
        - in: query
          name: payTo
          required: true
          schema: { type: string }
        - in: query
          name: minAmount
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerificationResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    Service:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        endpoint:
          type: string
        priceUsdc:
          type: number
        ownerAddress:
          type: string
        reputationScore:
          type: number
        totalCalls:
          type: integer
        avgRating:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - category
        - endpoint
        - priceUsdc
        - ownerAddress
        - reputationScore
        - totalCalls
        - avgRating
        - isActive
        - createdAt
    RegisterServiceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        endpoint:
          type: string
        priceUsdc:
          type: number
        ownerAddress:
          type: string
      required: [name, description, category, endpoint, priceUsdc, ownerAddress]
...
    PaymentRequiredResponse:
      type: object
      properties:
        error:
          type: string
        x402Version:
          type: integer
        accepts:
          type: array
          items:
            type: object
            properties:
              scheme:
                type: string
              network:
                type: string
              maxAmountRequired:
                type: string
              resource:
                type: string
              description:
                type: string
              mimeType:
                type: string
              payTo:
                type: string
              maxTimeoutSeconds:
                type: integer
              asset:
                type: string
              extra:
                type: object
            required: [scheme, network, maxAmountRequired, resource, payTo, asset]
      required: [error, x402Version, accepts]
...
    StellarBalanceResponse:
      type: object
      properties:
        address: { type: string }
        network: { type: string }
        balances:
          type: object
          properties:
            xlm: { type: string }
            usdc: { type: string }
          required: [xlm, usdc]
        horizonUrl: { type: string }
      required: [address, network, balances]
    BuildPaymentBody:
      type: object
      properties:
        fromSecretKey: { type: string }
        toAddress: { type: string }
        amountUsdc: { type: number }
        memo: { type: string }
      required: [fromSecretKey, toAddress, amountUsdc]
    BuildPaymentResponse:
      type: object
      properties:
        xdr: { type: string }
        fromAddress: { type: string }
        toAddress: { type: string }
        amountUsdc: { type: number }
        mppSplit:
          type: object
          properties:
            serviceAmount: { type: string }
            protocolFee: { type: string }
            protocolFeeAddress: { type: string }
          required: [serviceAmount, protocolFee, protocolFeeAddress]
        network: { type: string }
        instructions: { type: string }
      required: [xdr, fromAddress, toAddress, amountUsdc, mppSplit, network]
    SubmitPaymentResponse:
      type: object
      properties:
        txHash: { type: string }
        ledger: { type: integer }
        successful: { type: boolean }
        explorerUrl: { type: string }
        nextStep: { type: string }
      required: [txHash, successful]
    PaymentVerificationResponse:
      type: object
      properties:
        valid: { type: boolean }
        txHash: { type: string }
        fromAddress: { type: string }
        toAddress: { type: string }
        amount: { type: string }
        assetCode: { type: string }
        memo: { type: string }
        error: { type: string }
        explorerUrl: { type: string }
      required: [valid, txHash]