# lib/api-zod/src/generated/types/buildPaymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stellar Agent Trust Fabric API
 * OpenAPI spec version: 0.1.0
 */
import type { BuildPaymentResponseMppSplit } from "./buildPaymentResponseMppSplit";

export interface BuildPaymentResponse {
  xdr: string;
  fromAddress: string;
  toAddress: string;
  amountUsdc: number;
  mppSplit: BuildPaymentResponseMppSplit;
  network: string;
  instructions?: string;
}