# lib/api-zod/src/generated/types/paymentVerificationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stellar Agent Trust Fabric API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentVerificationResponse {
  valid: boolean;
  txHash: string;
  fromAddress?: string;
  toAddress?: string;
  amount?: string;
  assetCode?: string;
  memo?: string;
  error?: string;
  explorerUrl?: string;
}