# lib/api-zod/src/generated/types/stellarBalanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stellar Agent Trust Fabric API
 * OpenAPI spec version: 0.1.0
 */
import type { StellarBalanceResponseBalances } from "./stellarBalanceResponseBalances";

export interface StellarBalanceResponse {
  address: string;
  network: string;
  balances: StellarBalanceResponseBalances;
  horizonUrl?: string;
}