# lib/api-zod/src/generated/types/submitPaymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stellar Agent Trust Fabric API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitPaymentResponse {
  txHash: string;
  ledger?: number;
  successful: boolean;
  explorerUrl?: string;
  nextStep?: string;
}